Add following users
Users should be able to follow others and be followed by others. These
numbers should be displayed alongside their name and profile.

As a social network we've recognised that the experience for our users
is going to be greatly enhanced if we build a personalised arrogation of
posts relevant to their interests. Following and building a graph of
relationships is how we'll build this tailored user experience.